Module re_log_encoding::codec::arrow

Structs

Payload 🔒

Constants

DESERIALIZE_WITH_ARROW_1 🔒

SERIALIZE_WITH_ARROW_1 🔒

TODO(#3741): switch to arrow1 once https://github.com/apache/arrow-rs/issues/6803 is released

Functions

decode_arrow 🔒

encode_arrow 🔒

read_arrow_from_bytes 🔒

Helper function that deserializes raw bytes into arrow schema and record batch using Arrow IPC format.

write_arrow_to_bytes 🔒

Helper function that serializes given arrow schema and record batch into bytes using Arrow IPC format.