re_log_encoding::codec
Module arrow

Structs

Payload 🔒

Functions

decode_arrow 🔒

encode_arrow 🔒

read_arrow_from_bytes 🔒

Helper function that deserializes raw bytes into arrow schema and record batch using Arrow IPC format.

write_arrow_to_bytes 🔒

Helper function that serializes given arrow schema and record batch into bytes using Arrow IPC format.